third_party/blink/web_tests/external/wpt/css/css-overflow/scroll-markers/scroll-marker-next-focus-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Test: next focus search start after ::scroll-marker activation is from its ultimate originating element
  assert_equals: the first ::scroll-marker gets activated upon clicking expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
